# apps/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Help Me Get Hired API",
    "version": "0.0.0",
    "description": "The routes the web app calls on behalf of a Candidate. Every schema is generated from `packages/shared`, so the document and the code cannot disagree. Every route except `/health` needs the Session bearer token Auth.js issues in the web app; paste the value of its `session` cookie."
  },
  "servers": [
    {
      "url": "/",
      "description": "The API that serves this document; on the Docker Compose stack, http://localhost:API_PORT"
    }
  ],
...
  "security": [
    {
      "session": []
    }
  ],
  "paths": {
...
    "/profile/curation": {
      "get": {
        "tags": [
          "Curation"
        ],
        "summary": "The progress of the current Curation",
        "description": "The newest Curation of the Profile on screen: its status, the percentage of units saved, every unit with its kind, title, and status, the metrics counted from the Profile, the Model it runs on, and the failure reason with `resumeAfter` when it waits out a rate limit. The percentage counts saved units only, so every process answers the same number. An Account with no Curation for its latest Profile answers `progress: null`. Statements are not part of this answer. The `ETag` changes with any of it; send it back as `If-None-Match` to poll cheaply.",
        "operationId": "getCurationProgress",
        "parameters": [
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "description": "The `ETag` of the last answer",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The progress, or none yet",
            "headers": {
              "ETag": {
                "description": "Changes with anything in the answer",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationProgressState"
                }
              }
            }
          },
          "304": {
            "description": "Nothing changed since the `If-None-Match` value"
          },
          "401": {
            "description": "No valid Session bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/profile/curation/cancel": {
      "post": {
        "tags": [
          "Curation"
        ],
        "summary": "Stop the Curation",
        "description": "Moves the queued or running Curation to `cancelled`. The runner stops at its next unit boundary, the Statements already saved are kept for a retry, and nothing is indexed. Answers the progress as it now stands.",
        "operationId": "cancelCuration",
        "responses": {
          "200": {
            "description": "The progress, now cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationProgressState"
                }
              }
            }
          },
          "401": {
            "description": "No valid Session bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "`curation_not_found` when no Curation is queued or running",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiError"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "curation_not_found"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/profile/curation/retry": {
      "post": {
        "tags": [
          "Curation"
        ],
        "summary": "Try the failed or cancelled Curation again",
        "description": "Takes the newest Curation of the Profile, when it is `failed` or `cancelled`, back to `queued` with its attempts reset and its reason cleared, and adds its job again. Saved units stay saved, so the runner resumes at the first unsaved one and the Candidate pays only for what is left.",
        "operationId": "retryCuration",
        "responses": {
          "202": {
            "description": "The progress, queued again",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationProgressState"
                }
              }
            }
          },
          "401": {
            "description": "No valid Session bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "`curation_not_found` when the Profile has no Curation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiError"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "curation_not_found"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "`curation_active` while a Curation is queued or running",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiError"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "curation_active"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "`curation_not_ready` before the Profile is confirmed and a Model Key stored; `curation_not_retryable` when the Curation is not failed or cancelled; `curation_model_changed` when the Model Choice changed since it started, which a re-run answers instead",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiError"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "curation_not_ready",
                            "curation_not_retryable",
                            "curation_model_changed"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/profile/curation/rerun": {
      "post": {
        "tags": [
          "Curation"
        ],
        "summary": "Run the Curation again from zero",
        "description": "Creates a new Curation beside the current completed one, which stays current and retrievable until the new one completes and supersedes it; a re-run that fails or is cancelled leaves it intact. It spends the Candidate's tokens, so it is allowed only when the newest Curation of the Profile failed or was cancelled, or the Profile, the Model, or the prompt version changed since the current one was produced.",
        "operationId": "rerunCuration",
        "responses": {
          "202": {
            "description": "The progress of the new Curation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurationProgressState"
                }
              }
            }
          },
          "401": {
            "description": "No valid Session bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "`curation_active` while a Curation is queued or running",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiError"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "curation_active"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "`curation_not_ready` before the Profile is confirmed and a Model Key stored; `curation_unchanged` when running it again would produce the same Statements",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiError"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "curation_not_ready",
                            "curation_unchanged"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    }
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "issues": {
            "maxItems": 1000,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "statusCode",
          "message"
        ],
        "additionalProperties": false
      },
...
      "CurationProgressState": {
        "type": "object",
        "properties": {
          "progress": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "curationId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "queued",
                      "running",
                      "completed",
                      "failed",
                      "cancelled",
                      "superseded"
                    ]
                  },
                  "percentage": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 100
                  },
                  "units": {
                    "type": "object",
                    "properties": {
                      "total": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 9007199254740991
                      },
                      "saved": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 9007199254740991
                      },
                      "list": {
                        "maxItems": 1000,
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "format": "uuid",
                              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                            },
                            "kind": {
                              "type": "string",
                              "enum": [
                                "experience",
                                "project",
                                "cross_cutting",
                                "synthesis"
                              ]
                            },
                            "title": {
                              "type": "string",
                              "minLength": 1
                            },
                            "status": {
                              "type": "string",
                              "enum": [
                                "pending",
                                "running",
                                "saved",
                                "failed"
                              ]
                            },
                            "failureReason": {
                              "anyOf": [
                                {
                                  "type": "string",
                                  "enum": [
                                    "timeout",
                                    "invalid_output",
                                    "truncated_response",
                                    "provider_error"
                                  ]
                                },
                                {
                                  "type": "null"
                                }
                              ]
                            }
                          },
                          "required": [
                            "id",
                            "kind",
                            "title",
                            "status",
                            "failureReason"
                          ],
                          "additionalProperties": false
                        }
                      }
                    },
                    "required": [
                      "total",
                      "saved",
                      "list"
                    ],
                    "additionalProperties": false
                  },
                  "metrics": {
                    "type": "object",
                    "properties": {
                      "careerDuration": {
                        "type": "object",
                        "properties": {
                          "years": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991
                          },
                          "months": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 11
                          }
                        },
                        "required": [
                          "years",
                          "months"
                        ],
                        "additionalProperties": false
                      },
                      "durationPerCompany": {
                        "maxItems": 1000,
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "company": {
                              "type": "string",
                              "minLength": 1
                            },
                            "duration": {
                              "type": "object",
                              "properties": {
                                "years": {
                                  "type": "integer",
                                  "minimum": 0,
                                  "maximum": 9007199254740991
                                },
                                "months": {
                                  "type": "integer",
                                  "minimum": 0,
                                  "maximum": 11
                                }
                              },
                              "required": [
                                "years",
                                "months"
                              ],
                              "additionalProperties": false
                            }
                          },
                          "required": [
                            "company",
                            "duration"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "counts": {
                        "type": "object",
                        "properties": {
                          "roles": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991
                          },
                          "projects": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991
                          },
                          "certifications": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991
                          },
                          "languages": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991
                          },
                          "education": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 9007199254740991
                          }
                        },
                        "required": [
                          "roles",
                          "projects",
                          "certifications",
                          "languages",
                          "education"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "required": [
                      "careerDuration",
                      "durationPerCompany",
                      "counts"
                    ],
                    "additionalProperties": false
                  },
                  "modelId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "failureReason": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "attempts_exhausted",
                          "model_key_rejected"
                        ]
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "resumeAfter": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "pauseReason": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "provider_rate_limit",
                          "embedding_ceiling"
                        ]
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "rerun": {
                    "type": "object",
                    "properties": {
                      "allowed": {
                        "type": "boolean"
                      },
                      "refusal": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "curation_active",
                              "curation_not_ready",
                              "curation_unchanged"
                            ]
                          },
                          {
                            "type": "null"
                          }
                        ]
                      }
                    },
                    "required": [
                      "allowed",
                      "refusal"
                    ],
                    "additionalProperties": false
                  }
                },
                "required": [
                  "curationId",
                  "status",
                  "percentage",
                  "units",
                  "metrics",
                  "modelId",
                  "failureReason",
                  "resumeAfter",
                  "pauseReason",
                  "rerun"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "progress"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "http",
        "scheme": "bearer",
        "description": "The Session token the web app keeps in its `session` cookie after sign in"
      },
      "modelKeyTicket": {
        "type": "http",
        "scheme": "bearer",
        "description": "A single-use Model Key ticket from `POST /account/model/key-ticket`; only `PUT /account/model` accepts it"
      }
    }
  }
}